backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
django-filter==23.1
Pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
django-cloudinary-storage==0.3.0
cloudinary==1.32.0
stripe==5.5.0
celery==5.2.7
redis==4.5.4
django-redis==5.2.0
drf-yasg==1.21.7
django-debug-toolbar==4.0.0
